Split libnewsboat-ffi into modules
